Rearranged files for better gamefeel
